refactor to common
